ya esta el reporte de cada una de las mesas

+++ miProyecto/Oasis/models.py
@@ -160,7 +160,7 @@ class Mesa(models.Model):
     codigo_qr = models.CharField(max_length=100, unique=True)
     qr_imagen = models.ImageField(upload_to='mesas_qr_codes/', blank=True, null=True)
     usuario = models.ForeignKey('Usuario', on_delete=models.CASCADE, null=True, blank=True, default=None)
-    total_ganancia = models.DecimalField(max_digits=10, decimal_places=2, default=0)  # Campo para almacenar ganancias
+    total_ganancia = models.IntegerField(default=0)
 
     def __str__(self):
         return f'{self.id}'
